+++ Source/SMStdioOut.h
@@ -18,7 +18,7 @@ class StdioOut : public GenericOut {
 
     ~StdioOut() override;
     int put_byte(int c) override;
-    size_t write(const void* ptr, size_t size) override;
+    int write(const void* ptr, size_t size) override;
 };
 
 } // namespace SpectMorph
